Edit migration problem with accounts.id key

+++ database/migrations/2019_10_20_102009_create_profiles_table.php
@@ -15,7 +15,7 @@ class CreateProfilesTable extends Migration
     {
         Schema::create('cp_profiles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('account_id');
+            $table->unsignedInteger('account_id');
             $table->string('avatar')->nullable();
             $table->string('timezone')->nullable();
             $table->timestamps();
