+++ database/migrations/2019_10_23_090837_create_bugs_logs_table.php
@@ -17,7 +17,7 @@ class CreateBugsLogsTable extends Migration
             $table->bigIncrements('id');
 
             $table->unsignedBigInteger('bug_id');
-            $table->unsignedBigInteger('account_id');
+            $table->unsignedInteger('account_id');
 
             $table->string('logs', 250);
